# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Insurance Sale System APIs",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login & return token",
                "operationId": "a80bd0b5a103aa850a69329b0c482162",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string"
                                        },
                                        "expires_in": {
                                            "description": "number in seconds",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/me": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "user profile",
                "operationId": "439cabd18d601137ca7ea0567f0c64a3",
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "number"
                                                },
                                                "type": {
                                                    "type": "string"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "email_verified_at": {
                                                    "type": "datetime"
                                                },
                                                "password_changed_at": {
                                                    "type": "datetime"
                                                },
                                                "active": {
                                                    "type": "boolean"
                                                },
                                                "timezone": {
                                                    "type": "string"
                                                },
                                                "last_login_at": {
                                                    "type": "datetime"
                                                },
                                                "last_login_ip": {
                                                    "type": "string"
                                                },
                                                "to_be_logged_out": {
                                                    "type": "boolean"
                                                },
                                                "provider": {
                                                    "type": "string"
                                                },
                                                "provider_at": {
                                                    "type": "datetime"
                                                },
                                                "created_at": {
                                                    "type": "datetime"
                                                },
                                                "updated_at": {
                                                    "type": "datetime"
                                                },
                                                "deleted_at": {
                                                    "type": "datetime"
                                                },
                                                "avatar": {
                                                    "type": "string"
                                                },
                                                "permissions": {
                                                    "type": "object"
                                                },
                                                "roles": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout from Profile",
                "operationId": "b2471a52aab5ab098565d71a02713468",
                "responses": {
                    "200": {
                        "description": "Successfully logged out",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "refresh token",
                "operationId": "dc0f18b12412d6e8e0c2a9cbb4fec387",
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string"
                                        },
                                        "expires_in": {
                                            "description": "number in seconds",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/insurance-company": {
            "get": {
                "tags": [
                    "Insurance Companies"
                ],
                "summary": "Get list of Insurance Companies",
                "description": "Returns list of Insurance Companies",
                "operationId": "getInsuranceCompanyList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InsuranceCompanyResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/insurance-company/{companyId}/delete": {
            "delete": {
                "tags": [
                    "Insurance Companies"
                ],
                "summary": "Delete existing Insurance Company",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteInsuranceCompany",
                "parameters": [
                    {
                        "name": "companyId",
                        "in": "path",
                        "description": "company id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/insurance": {
            "get": {
                "tags": [
                    "Insurances"
                ],
                "summary": "Get list of Insurances",
                "description": "Returns list of Insurances",
                "operationId": "getInsuranceList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InsuranceResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/insurance/{insuranceId}/delete": {
            "delete": {
                "tags": [
                    "Insurances"
                ],
                "summary": "Delete existing Insurance",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteInsurance",
                "parameters": [
                    {
                        "name": "insuranceId",
                        "in": "path",
                        "description": "insurance id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Insurance": {
                "title": "Insurance",
                "description": "Insurance model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "title": "title",
                        "description": "title of the new insurance company",
                        "type": "string",
                        "example": "Iran Insurance"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description of the new insurance company",
                        "type": "string",
                        "example": "This is new insurance company's description"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "deleted_at": {
                        "title": "Deleted at",
                        "description": "Deleted at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "created_by": {
                        "title": "creator ID",
                        "description": "Creator's id of the new Insurance Company",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Insurance"
                }
            },
            "InsuranceCompany": {
                "title": "Insurance Company",
                "description": "Insurance Company model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "title": "title",
                        "description": "title of the new insurance company",
                        "type": "string",
                        "example": "Iran Insurance"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description of the new insurance company",
                        "type": "string",
                        "example": "This is new insurance company's description"
                    },
                    "website": {
                        "title": "website",
                        "description": "website of the new insurance company",
                        "type": "string",
                        "example": "https://iraninsurance.ir"
                    },
                    "logo": {
                        "title": "logo",
                        "description": "logo of the new insurance company",
                        "type": "string",
                        "example": "https://iraninsurance.ir/logo.png"
                    },
                    "bio": {
                        "title": "bio",
                        "description": "bio of the new insurance company",
                        "type": "string",
                        "example": "Iran Insurance bio text..."
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "deleted_at": {
                        "title": "Deleted at",
                        "description": "Deleted at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "created_by": {
                        "title": "creator ID",
                        "description": "Creator's id of the new Insurance Company",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    }
                },
                "type": "object",
                "xml": {
                    "name": "InsuranceCompany"
                }
            },
            "InsuranceCompanyResource": {
                "title": "InsuranceCompanyResource",
                "description": "Insurance Company resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/InsuranceCompany"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "InsuranceCompanyResource"
                }
            },
            "InsuranceResource": {
                "title": "InsuranceResource",
                "description": "Insurance resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Insurance"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "InsuranceResource"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
    "security": [
        []
    ]
}